RentSoft/WebServicespyRest/requirements.txt
astroid==2.3.3
cffi==1.13.2
Click==7.0
colorama==0.4.1
cryptography==2.8
Flask==1.1.1
Flask-Cors==3.0.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.17
pip==19.0.3
protobuf==3.9.2
pycparser==2.19
pycryptodome==3.9.4
PyJWT==1.7.1
pylint==2.4.3
setuptools==40.8.0
six==1.13.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
